+++ pyhive/__init__.py
@@ -2,4 +2,4 @@ from __future__ import absolute_import
 from __future__ import unicode_literals
 
 # major, minor, patch (upstream) and preset version
-__version__ = "0.7.0.3"
+__version__ = "0.7.0.4"